Change simple-cdd configuration to have full installer work offline
We move the default preseed to the kali profile so that the default
preseed can actually again enable simple-cdd-profiles that we need
because we want the other profiles to have their preseed and postinst
working (and they work through the script run when simple-cdd-profiles
is installed).

We add an offline profile that disables the network mirror (but not
the networking!) and we add a "postinst" script that is run in the
finish-install step to actually configure /etc/apt/sources.list that
is empty due to the preseed.